+++ tests/audio.rs
@@ -79,6 +79,12 @@ fn test_free_lossless_audio_codec() {
 }
 
 #[test]
+fn test_iff_8_bit_sampled_voice() {
+    let fmt = FileFormat::from_file("fixtures/audio/sample.8svx").unwrap();
+    assert_eq!(fmt, FileFormat::Iff8BitSampledVoice);
+}
+
+#[test]
 fn test_impulse_tracker_module() {
     let fmt = FileFormat::from_file("fixtures/audio/sample.it").unwrap();
     assert_eq!(fmt, FileFormat::ImpulseTrackerModule);